cdk/requirements.txt
aws-cdk-lib>=2.100.0
constructs>=10.0.0
aws-cdk-lib>=2.100.0
constructs>=10.0.0
aws-cdk.aws-lambda-python-alpha>=2.100.0a0
